# application/api-spec/src/main/resources/openapi/hotel-reservation.yaml
openapi: 3.0.3
info:
  title: Hotel Reservation API
  version: 1.0.0
  description: |
    Hotel Reservation backend API for room availability search, reservation management,
    and staff stay operations.
servers:
  - url: http://localhost:8080
paths:
...
  /admin/hotels:
    post:
      tags:
        - Admin Hotels
      operationId: createHotel
      summary: Create hotel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHotelRequest'
      responses:
        '201':
          description: Hotel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/hotels/{hotelId}:
    put:
      tags:
        - Admin Hotels
      operationId: updateHotel
      summary: Update hotel
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHotelRequest'
      responses:
        '200':
          description: Hotel updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Hotel not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/room-types:
    post:
      tags:
        - Admin Room Types
      operationId: createRoomType
      summary: Create room type
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomTypeRequest'
      responses:
        '201':
          description: Room type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomTypeResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Hotel not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/room-types/{roomTypeId}:
    put:
      tags:
        - Admin Room Types
      operationId: updateRoomType
      summary: Update room type
      security:
        - bearerAuth: []
      parameters:
        - name: roomTypeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomTypeRequest'
      responses:
        '200':
          description: Room type updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomTypeResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room type not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/rooms:
    post:
      tags:
        - Admin Rooms
      operationId: createRoom
      summary: Create room
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Hotel or room type not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/rooms/{roomId}:
    put:
      tags:
        - Admin Rooms
      operationId: updateRoom
      summary: Update room
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: Room updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room, hotel, or room type not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/hotels/{hotelId}/services:
    post:
      tags:
        - Admin Service Offerings
      operationId: createServiceOffering
      summary: Create hotel service offering
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceOfferingRequest'
      responses:
        '201':
          description: Service offering created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelServiceOfferingResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Hotel not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/hotels/{hotelId}/services/{serviceId}:
    put:
      tags:
        - Admin Service Offerings
      operationId: updateServiceOffering
      summary: Update hotel service offering
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceOfferingRequest'
      responses:
        '200':
          description: Service offering updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelServiceOfferingResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Hotel or service offering not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Admin Service Offerings
      operationId: deactivateServiceOffering
      summary: Deactivate hotel service offering
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Service offering deactivated
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Hotel or service offering not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HotelResponse:
      type: object
      required:
        - hotelId
        - name
        - city
        - country
        - address
        - stars
        - status
        - childrenAllowed
        - petsAllowed
        - infantMaxAge
        - childMaxAge
        - adultEquivalentAge
      properties:
        hotelId:
          type: integer
          format: int64
        name:
          type: string
        city:
          type: string
        country:
          type: string
        address:
          type: string
        stars:
          type: integer
          minimum: 1
          maximum: 5
        description:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, UNDER_MAINTENANCE]
        childrenAllowed:
          type: boolean
          description: Whether this hotel accepts children and infants. If false, childrenAges must be empty when searching or creating a reservation.
        petsAllowed:
          type: boolean
        infantMaxAge:
          type: integer
          description: Inclusive maximum age classified as an infant for this hotel. Infants are submitted inside childrenAges.
        childMaxAge:
          type: integer
          description: Inclusive maximum age that counts as a child for this hotel. Older minors from adultEquivalentAge count against adult occupancy instead.
        adultEquivalentAge:
          type: integer
          description: Inclusive age from which a minor submitted in childrenAges counts as an adult for occupancy limits. Must be less than 18.
    HotelServiceOfferingResponse:
      type: object
      required:
        - serviceOfferingId
        - hotelId
        - code
        - name
        - priceAmount
        - priceCurrency
        - active
      properties:
        serviceOfferingId:
          type: integer
          format: int64
        hotelId:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
        priceAmount:
          type: number
        priceCurrency:
          type: string
        active:
          type: boolean
        availabilityRule:
          type: string
    CreateHotelRequest:
      type: object
      required:
        - hotelId
        - name
        - city
        - country
        - address
        - stars
        - status
        - childrenAllowed
        - petsAllowed
        - infantMaxAge
        - childMaxAge
        - adultEquivalentAge
      properties:
        hotelId:
          type: integer
          format: int64
        name:
          type: string
        city:
          type: string
        country:
          type: string
        address:
          type: string
        stars:
          type: integer
          minimum: 1
          maximum: 5
        description:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, UNDER_MAINTENANCE]
        childrenAllowed:
          type: boolean
          description: Whether this hotel accepts children and infants. If false, childrenAges must be empty when searching or creating a reservation.
        petsAllowed:
          type: boolean
        infantMaxAge:
          type: integer
          minimum: 0
          description: Inclusive maximum age classified as an infant for this hotel. Infants are submitted inside childrenAges.
        childMaxAge:
          type: integer
          minimum: 0
          description: Inclusive maximum age that counts as a child for this hotel. Older minors from adultEquivalentAge count against adult occupancy instead.
        adultEquivalentAge:
          type: integer
          minimum: 1
          maximum: 17
          description: Inclusive age from which a minor submitted in childrenAges counts as an adult for occupancy limits. Must be greater than infantMaxAge.
    UpdateHotelRequest:
      type: object
      required:
        - name
        - city
        - country
        - address
        - stars
        - status
        - childrenAllowed
        - petsAllowed
        - infantMaxAge
        - childMaxAge
        - adultEquivalentAge
      properties:
        name:
          type: string
        city:
          type: string
        country:
          type: string
        address:
          type: string
        stars:
          type: integer
          minimum: 1
          maximum: 5
        description:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, UNDER_MAINTENANCE]
        childrenAllowed:
          type: boolean
          description: Whether this hotel accepts children and infants. If false, childrenAges must be empty when searching or creating a reservation.
        petsAllowed:
          type: boolean
        infantMaxAge:
          type: integer
          minimum: 0
          description: Inclusive maximum age classified as an infant for this hotel. Infants are submitted inside childrenAges.
        childMaxAge:
          type: integer
          minimum: 0
          description: Inclusive maximum age that counts as a child for this hotel. Older minors from adultEquivalentAge count against adult occupancy instead.
        adultEquivalentAge:
          type: integer
          minimum: 1
          maximum: 17
          description: Inclusive age from which a minor submitted in childrenAges counts as an adult for occupancy limits. Must be greater than infantMaxAge.
    RoomAmenityCode:
      type: string
      enum:
        - WIFI
        - AIR_CONDITIONING
        - HEATING
        - PRIVATE_BATHROOM
        - BATHTUB
        - SHOWER
        - HAIRDRYER
        - SAFE
        - MINIBAR
        - COFFEE_MACHINE
        - SMART_TV
        - SOUNDPROOFING
        - BALCONY
        - SEA_VIEW
        - CITY_VIEW
        - MOUNTAIN_VIEW
        - WORKSPACE
        - WHEELCHAIR_ACCESSIBLE
    RoomTypeResponse:
      type: object
      required:
        - roomTypeId
        - hotelId
        - name
        - maxAdults
        - maxChildren
        - maxInfants
        - maxTotalGuests
        - petsAllowed
        - maxPets
        - allowedPetTypes
        - basePriceAmount
        - basePriceCurrency
      properties:
        roomTypeId:
          type: integer
          format: int64
        hotelId:
          type: integer
          format: int64
        name:
          type: string
        maxAdults:
          type: integer
        maxChildren:
          type: integer
          description: Maximum number of non-infant children allowed for this room type.
        maxInfants:
          type: integer
          description: Maximum number of infants allowed for this room type. Infants are classified from childrenAges using the hotel's infantMaxAge.
        maxTotalGuests:
          type: integer
          description: Maximum occupancy counted as adults, adult-equivalent minors, and non-infant children. Infants are limited separately by maxInfants.
        petsAllowed:
          type: boolean
        maxPets:
          type: integer
        allowedPetTypes:
          type: array
          items:
            type: string
            enum: [DOG, CAT, OTHER]
        maxPetWeightKg:
          type: number
        petFeeAmount:
          type: number
        petFeeCurrency:
          type: string
        basePriceAmount:
          type: number
        basePriceCurrency:
          type: string
        description:
          type: string
        bedSetup:
          type: string
          nullable: true
          maxLength: 120
          description: Human-readable bed configuration as a single line. Null when not yet configured.
          example: 1 king bed + sofa bed
        roomSizeSqm:
          type: number
          format: float
          nullable: true
          minimum: 1
          maximum: 1000
          description: Room floor area in square metres. Null when unknown.
          example: 28.5
        amenities:
          type: array
          description: Room amenities as stable enum codes. Frontend maps each code to an icon and a localized label.
          items:
            $ref: '#/components/schemas/RoomAmenityCode'
          example: [WIFI, AIR_CONDITIONING, PRIVATE_BATHROOM, SEA_VIEW]
    CreateRoomTypeRequest:
      type: object
      required:
        - roomTypeId
        - hotelId
        - name
        - maxAdults
        - maxChildren
        - maxInfants
        - maxTotalGuests
        - petsAllowed
        - maxPets
        - allowedPetTypes
        - basePriceAmount
        - basePriceCurrency
      properties:
        roomTypeId:
          type: integer
          format: int64
        hotelId:
          type: integer
          format: int64
        name:
          type: string
        maxAdults:
          type: integer
          minimum: 1
        maxChildren:
          type: integer
          minimum: 0
          description: Maximum number of non-infant children allowed for this room type.
        maxInfants:
          type: integer
          minimum: 0
          description: Maximum number of infants allowed for this room type. Infants are classified from childrenAges using the hotel's infantMaxAge.
        maxTotalGuests:
          type: integer
          minimum: 1
          description: Maximum occupancy counted as adults, adult-equivalent minors, and non-infant children. Infants are limited separately by maxInfants.
        petsAllowed:
          type: boolean
        maxPets:
          type: integer
          minimum: 0
        allowedPetTypes:
          type: array
          items:
            type: string
            enum: [DOG, CAT, OTHER]
        maxPetWeightKg:
          type: number
          minimum: 0
        petFeeAmount:
          type: number
          minimum: 0
        petFeeCurrency:
          type: string
        basePriceAmount:
          type: number
          minimum: 0
        basePriceCurrency:
          type: string
        description:
          type: string
        bedSetup:
          type: string
          nullable: true
          maxLength: 120
          description: Human-readable bed configuration as a single line. Blank is treated as null.
          example: 1 king bed + sofa bed
        roomSizeSqm:
          type: number
          format: float
          nullable: true
          minimum: 1
          maximum: 1000
          description: Room floor area in square metres. Null when unknown.
          example: 28.5
        amenities:
          type: array
          description: Room amenities as stable enum codes. Empty array allowed.
          items:
            $ref: '#/components/schemas/RoomAmenityCode'
          example: [WIFI, AIR_CONDITIONING, PRIVATE_BATHROOM, SEA_VIEW]
    UpdateRoomTypeRequest:
      type: object
      required:
        - name
        - maxAdults
        - maxChildren
        - maxInfants
        - maxTotalGuests
        - petsAllowed
        - maxPets
        - allowedPetTypes
        - basePriceAmount
        - basePriceCurrency
      properties:
        name:
          type: string
        maxAdults:
          type: integer
          minimum: 1
        maxChildren:
          type: integer
          minimum: 0
          description: Maximum number of non-infant children allowed for this room type.
        maxInfants:
          type: integer
          minimum: 0
          description: Maximum number of infants allowed for this room type. Infants are classified from childrenAges using the hotel's infantMaxAge.
        maxTotalGuests:
          type: integer
          minimum: 1
          description: Maximum occupancy counted as adults, adult-equivalent minors, and non-infant children. Infants are limited separately by maxInfants.
        petsAllowed:
          type: boolean
        maxPets:
          type: integer
          minimum: 0
        allowedPetTypes:
          type: array
          items:
            type: string
            enum: [DOG, CAT, OTHER]
        maxPetWeightKg:
          type: number
          minimum: 0
        petFeeAmount:
          type: number
          minimum: 0
        petFeeCurrency:
          type: string
        basePriceAmount:
          type: number
          minimum: 0
        basePriceCurrency:
          type: string
        description:
          type: string
        bedSetup:
          type: string
          nullable: true
          maxLength: 120
          description: Human-readable bed configuration as a single line. Blank is treated as null.
          example: 1 king bed + sofa bed
        roomSizeSqm:
          type: number
          format: float
          nullable: true
          minimum: 1
          maximum: 1000
          description: Room floor area in square metres. Null when unknown.
          example: 28.5
        amenities:
          type: array
          description: Room amenities as stable enum codes. Empty array allowed.
          items:
            $ref: '#/components/schemas/RoomAmenityCode'
          example: [WIFI, AIR_CONDITIONING, PRIVATE_BATHROOM, SEA_VIEW]
    RoomResponse:
      type: object
      required:
        - roomId
        - hotelId
        - roomNumber
        - roomTypeId
        - capacity
        - status
      properties:
        roomId:
          type: integer
          format: int64
        hotelId:
          type: integer
          format: int64
        roomNumber:
          type: string
        roomTypeId:
          type: integer
          format: int64
        capacity:
          type: integer
        status:
          type: string
          enum: [AVAILABLE, CLEANING, MAINTENANCE, OUT_OF_SERVICE]
    CreateRoomRequest:
      type: object
      required:
        - roomId
        - hotelId
        - roomNumber
        - roomTypeId
        - capacity
        - status
      properties:
        roomId:
          type: integer
          format: int64
        hotelId:
          type: integer
          format: int64
        roomNumber:
          type: string
        roomTypeId:
          type: integer
          format: int64
        capacity:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [AVAILABLE, CLEANING, MAINTENANCE, OUT_OF_SERVICE]
    UpdateRoomRequest:
      type: object
      required:
        - hotelId
        - roomNumber
        - roomTypeId
        - capacity
        - status
      properties:
        hotelId:
          type: integer
          format: int64
        roomNumber:
          type: string
        roomTypeId:
          type: integer
          format: int64
        capacity:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [AVAILABLE, CLEANING, MAINTENANCE, OUT_OF_SERVICE]
    CreateServiceOfferingRequest:
      type: object
      required:
        - serviceOfferingId
        - code
        - name
        - priceAmount
        - priceCurrency
        - active
      properties:
        serviceOfferingId:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
        priceAmount:
          type: number
          minimum: 0
        priceCurrency:
          type: string
        active:
          type: boolean
        availabilityRule:
          type: string
    UpdateServiceOfferingRequest:
      type: object
      required:
        - code
        - name
        - priceAmount
        - priceCurrency
        - active
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        priceAmount:
          type: number
          minimum: 0
        priceCurrency:
          type: string
        active:
          type: boolean
        availabilityRule:
          type: string
...
    ErrorResponse:
      type: object
      description: RFC 7807 problem details response returned by the REST layer.
      properties:
        type:
          type: string
          example: about:blank
        title:
          type: string
          example: Validation failed
        status:
          type: integer
          format: int32
          example: 400
        detail:
          type: string
          example: checkOut must be after checkIn
        instance:
          type: string
          example: /reservations